Break down provider into multiple files
