+++ pkg/provider/containers.go
@@ -1,0 +1,74 @@
+// Copyright (C) 2022 Red Hat, Inc.
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+package provider
+
+import (
+	"errors"
+	"fmt"
+	"strings"
+
+	"github.com/sirupsen/logrus"
+	"github.com/test-network-function/cnf-certification-test/pkg/configuration"
+	corev1 "k8s.io/api/core/v1"
+)
+
+type Container struct {
+	*corev1.Container
+	Status                   corev1.ContainerStatus
+	Namespace                string
+	Podname                  string
+	NodeName                 string
+	Runtime                  string
+	UID                      string
+	ContainerImageIdentifier configuration.ContainerImageIdentifier
+}
+
+func GetContainer() *Container {
+	return &Container{}
+}
+
+func (c *Container) GetUID() (string, error) {
+	split := strings.Split(c.Status.ContainerID, "://")
+	uid := ""
+	if len(split) > 0 {
+		uid = split[len(split)-1]
+	}
+	if uid == "" {
+		logrus.Debugln(fmt.Sprintf("could not find uid of %s/%s/%s\n", c.Namespace, c.Podname, c.Name))
+		return "", errors.New("cannot determine container UID")
+	}
+	logrus.Debugln(fmt.Sprintf("uid of %s/%s/%s=%s\n", c.Namespace, c.Podname, c.Name, uid))
+	return uid, nil
+}
+
+func (c *Container) StringLong() string {
+	return fmt.Sprintf("node: %s ns: %s podName: %s containerName: %s containerUID: %s containerRuntime: %s",
+		c.NodeName,
+		c.Namespace,
+		c.Podname,
+		c.Name,
+		c.Status.ContainerID,
+		c.Runtime,
+	)
+}
+func (c *Container) String() string {
+	return fmt.Sprintf("container: %s pod: %s ns: %s",
+		c.Name,
+		c.Podname,
+		c.Namespace,
+	)
+}
